pyproject.toml
[project]
name = "langsmith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.10",
    "langchain-groq>=0.1.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "langsmith",
]

[tool.uv.sources]
langsmith = { workspace = true }
